[Silver II] Title: 나무 자르기, Time: 588 ms, Memory: 5928 KB -BaekjoonHub

+++ 백준/Silver/2805. 나무 자르기/나무 자르기.cc
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-long long tree[1000005]; // 나무
+int tree[1000005]; // 나무
 
 int n;
 long long m;
